IndexedDB/idbdatabase_deleteObjectStore5.htm

<!DOCTYPE html>
<title>IDBDatabase.deleteObjectStore() - object store 'name' set to null/undefined </title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="https://dvcs.w3.org/hg/IndexedDB/raw-file/tip/Overview.html#widl-IDBDatabase-deleteObjectStore-void-DOMString-name">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support.js"></script>
<div id="log"></div>
<script>
var t = async_test(),
    open_rq = createdb(t);

open_rq.onupgradeneeded = function(e) {
    var db = e.target.result;
    assert_throws('NotFoundError', function() { db.deleteObjectStore(null); });

    assert_throws('NotFoundError', function() { db.deleteObjectStore(undefined); });
    t.done();
}
</script>
